fix(@leav/ui): fix date filter not working options
